movies/pyproject.toml
[tool.poetry]
name = "new-admin-panel-sprint-1"
version = "0.1.0"
description = ""
authors = ["sgavrilenko <gryroach@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
psycopg = "^3.2.2"
faker = "^29.0.0"
python-dotenv = "^1.0.1"
isort = "^5.13.2"
sqlfluff = "^3.2.0"
flake8 = "^7.1.1"
django = "^5"
django-split-settings = "^1.3.2"
django-debug-toolbar = "^4.4.6"
python-dateutil = "^2.9.0.post0"
gunicorn = "^23.0.0"
gevent = "^24.2.1"
django-extensions = "^3.2.3"
django-cors-headers = "^4.5.0"
djangorestframework = "^3.15.2"
django-filter = "^24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
